requirements.txt
wheel
alembic==1.5.8
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
countryinfo==0.1.2
cycler==0.10.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pprintpp==0.4.0
psycopg2-binary
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
SQLAlchemy==1.4.7
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==2.0.1
